# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/perawat/pemeriksaan": {
            "get": {
                "tags": [
                    "Pemeriksaan Awal"
                ],
                "summary": "Ambil semua data pemeriksaan",
                "operationId": "2a4a2ad8da82e7a2f3ef28135d0a0d01",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pemeriksaan Awal"
                ],
                "summary": "Tambah pemeriksaan awal",
                "operationId": "6d7341d3dfe2520c4d5989b27d9a19bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ID_PEMERIKSAAN",
                                    "ID_DOKTER",
                                    "ID_PERAWAT",
                                    "ID_RESEP",
                                    "RM",
                                    "STATUS"
                                ],
                                "properties": {
                                    "ID_PEMERIKSAAN": {
                                        "type": "string",
                                        "example": "P001"
                                    },
                                    "ID_DOKTER": {
                                        "type": "string",
                                        "example": "D001"
                                    },
                                    "ID_PERAWAT": {
                                        "type": "string",
                                        "example": "P001"
                                    },
                                    "ID_RESEP": {
                                        "type": "string",
                                        "example": "R001"
                                    },
                                    "RM": {
                                        "type": "string",
                                        "example": "RM001"
                                    },
                                    "SUHU": {
                                        "type": "string",
                                        "example": "36"
                                    },
                                    "TENSI": {
                                        "type": "string",
                                        "example": "120/80"
                                    },
                                    "TINGGI_BADAN": {
                                        "type": "string",
                                        "example": "176"
                                    },
                                    "BERAT_BADAN": {
                                        "type": "string",
                                        "example": "70"
                                    },
                                    "STATUS": {
                                        "type": "string",
                                        "example": "Selesai"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data Terkirim"
                    },
                    "400": {
                        "description": "Data tidak valid"
                    }
                }
            }
        },
        "/api/perawat/pemeriksaan/{id}": {
            "get": {
                "tags": [
                    "Pemeriksaan Awal"
                ],
                "summary": "Ambil data pemeriksaan berdasarkan ID",
                "operationId": "489be6ff3f7364fc5f3d948944644185",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pemeriksaan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "P001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data ditemukan"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pemeriksaan Awal"
                ],
                "summary": "Update data pemeriksaan Awal",
                "operationId": "09d14feca756f983d0562edc7052beab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pemeriksaan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "P001"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ID_PERAWAT": {
                                        "type": "string",
                                        "example": "D002"
                                    },
                                    "SUHU": {
                                        "type": "string",
                                        "example": "45"
                                    },
                                    "TENSI": {
                                        "type": "string",
                                        "example": "70/120"
                                    },
                                    "BERAT_BADAN": {
                                        "type": "string",
                                        "example": "70"
                                    },
                                    "TINGGI_BADAN": {
                                        "type": "string",
                                        "example": "176"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diperbarui"
                    },
                    "400": {
                        "description": "Permintaan tidak valid"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pemeriksaan Awal"
                ],
                "summary": "Hapus data pemeriksaan berdasarkan ID",
                "operationId": "ebb0054c586c39595e7552deec33228c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pemeriksaan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "P001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        }
    },
...
}